+++ ss13_tools/centcom/__version__.py
@@ -1,0 +1,1 @@
+__version__ = '0.1.0'
