+++ ss13_tools/slur_detector/__main__.py
