+++ ss13_tools/slur_detector/constants.py
